requirements.txt
numpy
google-generativeai
python-dotenv